Moving react-next Shimmer to react-internal.
