--- packages/react-examples/src/react-next/Shimmer/docs/ShimmerDonts.md
@@ -1,4 +1,0 @@
-- Use on the same element both types of widths. It will always default to just one of them. See documentation below.
-- Build Shimmer UI should with a lot of details. Circles and rectangles are really as detailed as you want to get. Adding more detail will result in confusion once the UI loads.
-- Use shimmer if you are confident that the UI will take less than a second to load.
-- Use shimmer as a way to not make improvements in your code to improve performance.
