+++ packages/react-internal/src/components/Shimmer/Shimmer.base.tsx
@@ -1,92 +1,35 @@
 import * as React from 'react';
-import {
-  classNamesFunction,
-  DelayedRender,
-  getNativeProps,
-  divProperties,
-  Async,
-  initializeComponentRef,
-} from '../../Utilities';
+import { classNamesFunction, DelayedRender, getNativeProps, divProperties } from '../../Utilities';
 import { IShimmerProps, IShimmerStyleProps, IShimmerStyles } from './Shimmer.types';
 import { ShimmerElementsGroup } from './ShimmerElementsGroup/ShimmerElementsGroup';
-
-export interface IShimmerState {
-  /**
-   * Flag for knowing when to remove the shimmerWrapper from the DOM.
-   */
-  contentLoaded?: boolean;
-}
+import { useSetTimeout, useConst } from '@uifabric/react-hooks';
 
 const TRANSITION_ANIMATION_INTERVAL = 200; /* ms */
+const COMPONENT_NAME = 'Shimmer';
 
 const getClassNames = classNamesFunction<IShimmerStyleProps, IShimmerStyles>();
 
 /**
  * {@docCategory Shimmer}
  */
-export class ShimmerBase extends React.Component<IShimmerProps, IShimmerState> {
-  public static defaultProps: IShimmerProps = {
-    isDataLoaded: false,
-  };
-
-  private _classNames: { [key in keyof IShimmerStyles]: string };
-  private _lastTimeoutId: number;
-  private _async: Async;
-
-  constructor(props: IShimmerProps) {
-    super(props);
-
-    initializeComponentRef(this);
-
-    this.state = {
-      contentLoaded: props.isDataLoaded,
-    };
-
-    this._async = new Async(this);
-  }
-
-  public componentDidUpdate(prevProps: IShimmerProps): void {
-    const { isDataLoaded } = this.props;
-
-    if (isDataLoaded !== prevProps.isDataLoaded) {
-      this._async.clearTimeout(this._lastTimeoutId);
-
-      // Removing the shimmerWrapper div from the DOM only after the fade out animation completed.
-      if (isDataLoaded) {
-        this._lastTimeoutId = this._async.setTimeout(() => {
-          this.setState({
-            contentLoaded: isDataLoaded,
-          });
-        }, TRANSITION_ANIMATION_INTERVAL);
-      } else {
-        this.setState({
-          contentLoaded: isDataLoaded,
-        });
-      }
-    }
-  }
-
-  public componentWillUnmount(): void {
-    this._async.dispose();
-  }
-
-  public render(): JSX.Element {
+export const ShimmerBase: React.FunctionComponent<IShimmerProps> = React.forwardRef<HTMLDivElement, IShimmerProps>(
+  (props, ref) => {
     const {
       styles,
       shimmerElements,
       children,
-      isDataLoaded,
       width,
       className,
       customElementsGroup,
       theme,
       ariaLabel,
       shimmerColors,
-    } = this.props;
+      isDataLoaded = false,
+    } = props;
 
-    const { contentLoaded } = this.state;
+    const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(props, divProperties);
 
-    this._classNames = getClassNames(styles!, {
+    const classNames: { [key in keyof IShimmerStyles]: string } = getClassNames(styles!, {
       theme: theme!,
       isDataLoaded,
       className,
@@ -95,13 +38,39 @@ export class ShimmerBase extends React.Component<IShimmerProps, IShimmerState> {
       shimmerWaveColor: shimmerColors && shimmerColors.shimmerWave,
     });
 
-    const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(this.props, divProperties);
+    const internalState = useConst({
+      lastTimeoutId: 0,
+    });
+
+    const { setTimeout, clearTimeout } = useSetTimeout();
+
+    /**
+     * Flag for knowing when to remove the shimmerWrapper from the DOM.
+     */
+    const [contentLoaded, setContentLoaded] = React.useState(isDataLoaded);
+
+    const divStyleProp = { width: width ? width : '100%' };
+
+    React.useEffect(() => {
+      if (isDataLoaded !== contentLoaded) {
+        if (isDataLoaded) {
+          internalState.lastTimeoutId = setTimeout(() => {
+            setContentLoaded(true);
+          }, TRANSITION_ANIMATION_INTERVAL);
+
+          return () => clearTimeout(internalState.lastTimeoutId);
+        } else {
+          setContentLoaded(false);
+        }
+      }
+      // eslint-disable-next-line react-hooks/exhaustive-deps -- Should only run when isDataLoaded changes.
+    }, [isDataLoaded]);
 
     return (
-      <div {...divProps} className={this._classNames.root}>
+      <div {...divProps} className={classNames.root} ref={ref}>
         {!contentLoaded && (
-          <div style={{ width: width ? width : '100%' }} className={this._classNames.shimmerWrapper}>
-            <div className={this._classNames.shimmerGradient} />
+          <div style={divStyleProp} className={classNames.shimmerWrapper}>
+            <div className={classNames.shimmerGradient} />
             {customElementsGroup ? (
               customElementsGroup
             ) : (
@@ -112,15 +81,16 @@ export class ShimmerBase extends React.Component<IShimmerProps, IShimmerState> {
             )}
           </div>
         )}
-        {children && <div className={this._classNames.dataWrapper}>{children}</div>}
+        {children && <div className={classNames.dataWrapper}>{children}</div>}
         {ariaLabel && !isDataLoaded && (
           <div role="status" aria-live="polite">
             <DelayedRender>
-              <div className={this._classNames.screenReaderText}>{ariaLabel}</div>
+              <div className={classNames.screenReaderText}>{ariaLabel}</div>
             </DelayedRender>
           </div>
         )}
       </div>
     );
-  }
-}
+  },
+);
+ShimmerBase.displayName = COMPONENT_NAME;
