+++ packages/react-next/etc/react-next.api.md
@@ -56,7 +56,6 @@ import { IPickerAriaIds } from '@fluentui/react/lib/Pickers';
 import { IPickerItem } from '@fluentui/react/lib/Pickers';
 import { IPickerItemProps } from '@fluentui/react/lib/Pickers';
 import { IPositionedData } from '@fluentui/react/lib/Positioning';
-import { IRawStyle } from '@fluentui/react/lib/Styling';
 import { IRectangle } from '@fluentui/react/lib/Utilities';
 import { IRefObject } from '@fluentui/react/lib/Utilities';
 import { IRenderFunction } from '@fluentui/react/lib/Utilities';
@@ -858,11 +857,6 @@ export interface IChoiceGroupStyles {
     root?: IStyle;
 }
 
-// @public
-export interface ICircle extends IShimmerElement {
-    height?: number;
-}
-
 // @public (undocumented)
 export interface ICoachmark {
     dismiss?: (ev?: Event | React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>) => void;
@@ -1483,12 +1477,6 @@ export interface IFocusTrapZoneProps extends React.HTMLAttributes<HTMLDivElement
     isClickableOutsideFocusTrap?: boolean;
 }
 
-// @public
-export interface IGap extends IShimmerElement {
-    height?: number;
-    width?: number | string;
-}
-
 export { IGenericItem }
 
 export { IInputProps }
@@ -1533,12 +1521,6 @@ export interface ILayerStyleProps {
 export interface ILayerStyles {
     content?: IStyle;
     root?: IStyle;
-}
-
-// @public
-export interface ILine extends IShimmerElement {
-    height?: number;
-    width?: number | string;
 }
 
 // @public (undocumented)
@@ -1948,165 +1930,6 @@ export interface ISelectedPeopleProps extends IBaseSelectedItemsListProps<IExten
 }
 
 // @public (undocumented)
-export interface IShimmerCircle {
-}
-
-// @public
-export interface IShimmerCircleProps extends React.AllHTMLAttributes<HTMLElement> {
-    // @deprecated
-    borderStyle?: IRawStyle;
-    componentRef?: IRefObject<IShimmerCircle>;
-    height?: number;
-    styles?: IStyleFunctionOrObject<IShimmerCircleStyleProps, IShimmerCircleStyles>;
-    theme?: ITheme;
-}
-
-// @public
-export type IShimmerCircleStyleProps = {
-    theme: ITheme;
-    height?: number;
-    borderStyle?: IRawStyle;
-};
-
-// @public
-export interface IShimmerCircleStyles {
-    root?: IStyle;
-    svg?: IStyle;
-}
-
-// @public
-export interface IShimmerColors {
-    background?: string;
-    shimmer?: string;
-    shimmerWave?: string;
-}
-
-// @public
-export interface IShimmerElement {
-    height?: number;
-    type: ShimmerElementType;
-    verticalAlign?: 'top' | 'center' | 'bottom';
-    width?: number | string;
-}
-
-// @public (undocumented)
-export interface IShimmerElementsGroup {
-}
-
-// @public
-export interface IShimmerElementsGroupProps extends React.AllHTMLAttributes<HTMLElement> {
-    backgroundColor?: string;
-    componentRef?: IRefObject<IShimmerElementsGroup>;
-    flexWrap?: boolean;
-    rowHeight?: number;
-    shimmerElements?: IShimmerElement[];
-    styles?: IStyleFunctionOrObject<IShimmerElementsGroupStyleProps, IShimmerElementsGroupStyles>;
-    theme?: ITheme;
-    width?: string;
-}
-
-// @public
-export interface IShimmerElementsGroupStyleProps {
-    flexWrap?: boolean;
-    theme: ITheme;
-}
-
-// @public
-export interface IShimmerElementsGroupStyles {
-    root?: IStyle;
-}
-
-// @public (undocumented)
-export interface IShimmerGap {
-}
-
-// @public
-export interface IShimmerGapProps extends React.AllHTMLAttributes<HTMLElement> {
-    // @deprecated
-    borderStyle?: IRawStyle;
-    componentRef?: IRefObject<IShimmerGap>;
-    height?: number;
-    styles?: IStyleFunctionOrObject<IShimmerGapStyleProps, IShimmerGapStyles>;
-    theme?: ITheme;
-    width?: number | string;
-}
-
-// @public
-export type IShimmerGapStyleProps = {
-    theme: ITheme;
-    height?: number;
-    borderStyle?: IRawStyle;
-};
-
-// @public
-export interface IShimmerGapStyles {
-    root?: IStyle;
-}
-
-// @public (undocumented)
-export interface IShimmerLine {
-}
-
-// @public
-export interface IShimmerLineProps extends React.AllHTMLAttributes<HTMLElement> {
-    // @deprecated
-    borderStyle?: IRawStyle;
-    componentRef?: IRefObject<IShimmerLine>;
-    height?: number;
-    styles?: IStyleFunctionOrObject<IShimmerLineStyleProps, IShimmerLineStyles>;
-    theme?: ITheme;
-    width?: number | string;
-}
-
-// @public
-export type IShimmerLineStyleProps = {
-    theme: ITheme;
-    height?: number;
-    borderStyle?: IRawStyle;
-};
-
-// @public
-export interface IShimmerLineStyles {
-    bottomLeftCorner?: IStyle;
-    bottomRightCorner?: IStyle;
-    root?: IStyle;
-    topLeftCorner?: IStyle;
-    topRightCorner?: IStyle;
-}
-
-// @public
-export interface IShimmerProps extends React.AllHTMLAttributes<HTMLElement>, React.RefAttributes<HTMLElement> {
-    ariaLabel?: string;
-    className?: string;
-    customElementsGroup?: React.ReactNode;
-    isDataLoaded?: boolean;
-    shimmerColors?: IShimmerColors;
-    shimmerElements?: IShimmerElement[];
-    styles?: IStyleFunctionOrObject<IShimmerStyleProps, IShimmerStyles>;
-    theme?: ITheme;
-    width?: number | string;
-}
-
-// @public
-export interface IShimmerStyleProps {
-    className?: string;
-    isDataLoaded?: boolean;
-    shimmerColor?: string;
-    shimmerWaveColor?: string;
-    theme: ITheme;
-    transitionAnimationInterval?: number;
-}
-
-// @public
-export interface IShimmerStyles {
-    dataWrapper?: IStyle;
-    root?: IStyle;
-    screenReaderText?: IStyle;
-    shimmerGradient?: IStyle;
-    shimmerWrapper?: IStyle;
-}
-
-// @public (undocumented)
 export interface ISpinButton {
     focus: () => void;
     value?: string;
@@ -2484,50 +2307,6 @@ export class SelectedPeopleList extends BasePeopleSelectedItemsList {
     protected renderItems: () => JSX.Element[];
 }
 
-// @public (undocumented)
-export const Shimmer: React.FunctionComponent<IShimmerProps>;
-
-// @public (undocumented)
-export const ShimmerBase: React.FunctionComponent<IShimmerProps>;
-
-// @public (undocumented)
-export const ShimmerCircle: React.FunctionComponent<IShimmerCircleProps>;
-
-// @public (undocumented)
-export const ShimmerCircleBase: React.FunctionComponent<IShimmerCircleProps>;
-
-// @public
-export enum ShimmerElementsDefaultHeights {
-    circle = 24,
-    gap = 16,
-    line = 16
-}
-
-// @public (undocumented)
-export const ShimmerElementsGroup: React.FunctionComponent<IShimmerElementsGroupProps>;
-
-// @public (undocumented)
-export const ShimmerElementsGroupBase: React.FunctionComponent<IShimmerElementsGroupProps>;
-
-// @public
-export enum ShimmerElementType {
-    circle = 2,
-    gap = 3,
-    line = 1
-}
-
-// @public (undocumented)
-export const ShimmerGap: React.FunctionComponent<IShimmerGapProps>;
-
-// @public (undocumented)
-export const ShimmerGapBase: React.FunctionComponent<IShimmerGapProps>;
-
-// @public (undocumented)
-export const ShimmerLine: React.FunctionComponent<IShimmerLineProps>;
-
-// @public (undocumented)
-export const ShimmerLineBase: React.FunctionComponent<IShimmerLineProps>;
-
 // @public
 export const SpinButton: React.FunctionComponent<ISpinButtonProps>;
 
@@ -2758,6 +2537,7 @@ export * from "@fluentui/react/lib/ScrollablePane";
 export * from "@fluentui/react/lib/SelectableOption";
 export * from "@fluentui/react/lib/Selection";
 export * from "@fluentui/react/lib/Separator";
+export * from "@fluentui/react/lib/Shimmer";
 export * from "@fluentui/react/lib/ShimmeredDetailsList";
 export * from "@fluentui/react/lib/Spinner";
 export * from "@fluentui/react/lib/Stack";
