--- packages/react-next/src/components/Shimmer/index.ts
@@ -1,15 +1,0 @@
-export * from './Shimmer';
-export * from './Shimmer.base';
-export * from './Shimmer.types';
-export * from './ShimmerLine/ShimmerLine';
-export * from './ShimmerLine/ShimmerLine.base';
-export * from './ShimmerLine/ShimmerLine.types';
-export * from './ShimmerCircle/ShimmerCircle';
-export * from './ShimmerCircle/ShimmerCircle.base';
-export * from './ShimmerCircle/ShimmerCircle.types';
-export * from './ShimmerGap/ShimmerGap';
-export * from './ShimmerGap/ShimmerGap.base';
-export * from './ShimmerGap/ShimmerGap.types';
-export * from './ShimmerElementsGroup/ShimmerElementsGroup';
-export * from './ShimmerElementsGroup/ShimmerElementsGroup.base';
-export * from './ShimmerElementsGroup/ShimmerElementsGroup.types';
